requirements.txt
pandas>=2.3,<3
SQLAlchemy>=2.0,<3
psycopg2-binary>=2.9,<3
python-dotenv>=1.0
requests>=2.32
